pyproject.toml
[project]
name = "CIVISAIPolicyAnalyserBackend"
version = "1.0.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.11.3",
    "fastapi==0.115.11",
    "fastapi-cli==0.0.7",
    "FastAPI-SQLAlchemy==0.2.1",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "uvicorn>=0.34.2",
    "loguru>=0.7.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "coverage==7.8.0",
    "pytest-cov==6.1.1",
    "ruff>=0.11.6",
]

[tool.ruff.lint]
ignore = [
    "E501"  # Line is Too Long
]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
